Fixed mistake in AD1PCFG setting. Added EEPROM reset function.
To reset EEPROM to factory defaults, short pin RB15 with ground and apply power to the board (On the revision 3 baseboard this is achieved by shorting Pin 2 and Pin 3 on header P2).

The board will blink the orange frame LED 10 times to indicate a successful wipe. The error LED will then turn on because the configuration memory CRC has been invalidated. Proceed to program the board as normal.